+++ routes/api.php
@@ -23,10 +23,9 @@ Route::group(['middleware' => 'auth:api'], function () {
     Route::resource('offer', 'Api\OfferController')->except(['index', 'show']);
     Route::resource('institution', 'Api\InstitutionController')->except(['index', 'show']);
     Route::get('subscription', 'Api\SubscriptionController@index'); # get all subscriptions
-    Route::get('subscription/user/{user_id}', 'Api\SubscriptionController@indexOffersForUser'); # get all subscriptions of user
-    Route::get('subscription/offer/{offer_id}', 'Api\SubscriptionController@indexUsersForOffer'); # get all subscriptions of offer
-    Route::get('subscription/{offer_id}/{user_id}', 'Api\SubscriptionController@showFromIds'); # get specific subscription from user/offer ids
-    Route::resource('subscription', 'Api\SubscriptionController');
+    Route::get('subscription/user/{user_id}', 'Api\UserController@offers'); # get all subscriptions of user
+    Route::get('subscription/offer/{offer_id}', 'Api\OfferController@users'); # get all subscriptions of offer
+    Route::get('subscription/{user_id}/{offer_id}', 'Api\OfferController@subscription'); # get a specific subscription
     Route::get('user/email', 'Api\UserController@showFromEmail'); # get user by email
     Route::resource('user', 'Api\UserController');
 });